created migration files. truncated the database

+++ database/migrations/2026_04_18_200002_create_payments_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
 
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('booking_id');
+            $table->unsignedInteger('booking_id');
             $table->decimal('total_amount', 10, 2);
             $table->decimal('amount_tendered', 10, 2)->nullable();
             $table->decimal('change_amount', 10, 2)->nullable();
